Added alignDepth plotting functions.
Added a set of functions to plot alignDepths and gene structures
together with each other.

plotTranscripts
pu.arrows (plot unit arrows)
plotDepthsTranscripts

These should not be considered as finished products. In particular the
plotTranscripts function depends on a specific list / dataframe
structure, that isn't formally described anywhere. I've used this as it
interfaces easily with a function I have for pulling gene transcripts
from a connection to an ensembl database. But it's difficult to make a
formal dependency on those functions. (Though I suppose one could access
directly a publich Ensembl database?).