Temporary modification to limit the bam_fetch function
Modified the bam_fetch function to restrict the matches it fetches to
those which both begin and end within the specified range. This, however,
does not seem to be the best way to achieve the current aim (to distinguish
transgene from native gene expression) for the following reasons.

1. It modifies what appears to be a fairly core function of the samtools library for
which there may be a large number of dependancies.

2. It relies on parsing the CIGAR string to find the end of the alignment. However, this
information must already be present within the index and it should be possible to obtain it
much more rapidly by modifying the bam_iter_query() function to take a flag of some sort.
However, I'm having trouble understanding how the bam_iter_query function works.

3. The function needs to be rewritten to take a specific flag indicating what kinds of alignments
are desired. We can then also specify other alignment flags (eg. the strand) that are
more generally useful.

So to overcome this I will:

1. rename the current function and change it's protototype to take 2 flag arguments.
2. introduce a new function using the old name and prototype that simply calls the
renamed function.